5475: Refactor Aggregation Filter Iterator
