+++ java/core/src/main/java/sleeper/core/iterator/AggregatorIteratorImpl.java
@@ -15,8 +15,6 @@
  */
 package sleeper.core.iterator;
 
-import sleeper.core.iterator.AggregationFilteringIterator.Aggregation;
-import sleeper.core.iterator.AggregationFilteringIterator.FilterAggregationConfig;
 import sleeper.core.key.Key;
 import sleeper.core.row.Row;
